Factor mdbook preprocessor into library for use in mdbook-quiz. Fix rendering issues with leader-line.
